requirements.txt
streamlit
nbdt==0.0.4
numpy
torch==1.8.1
torchvision==0.9.1
pandas
plotly.express
streamlit-aggrid
openpyxl
pyxlsb
matplotlib
wordcloud
